Added font check

+++ install.ps1
@@ -119,7 +119,9 @@ if (-not (Get-Command gcc -errorAction silentlyContinue)) {
   scoop install gcc
 }
 
-oh-my-posh font install Meslo
+if (-not (Test-Path $env:SystemRoot\Fonts\Meslo* -PathType Leaf)) {
+  oh-my-posh font install Meslo
+}
 
 Write-Host "Completed Windows install script"
